Improve how to publish documentation.
See merge request meetupbot/mitup-telegram-bot!75

- Approved by: Approved-by: Enrique Araque <minixute@gmail.com>
- Merged by: Enrique Araque <minixute@gmail.com>

Commits squashed (max 100): 

* Improve how to publish documentation.

Create a command to publish the documentation that after syncing with
s3, gets the list of files that have been updated and expires them from
the CloudFront cache.

This is needed because CloudFront has a TTL of 24h. While we can reduce
this TTL most of the time we don't need CloudFront to expire the
objects. In this case it is better to expire them manually and only run
the jobs when any documentation file has changed.